refactor perforce API to modules for similar commands
